Forgot to add the rest of the code to previous commit

+++ include/other/parse_ports.h
@@ -10,4 +10,4 @@
         Length of array.
 
 */
-int parse_ports(char * str, int ** ports);
+int parse_ports(char * str, unsigned short ** ports);
